Migrate service health check to Go
Fix #3

Replace the existing service checking functionality with a Go
application. Previously the logic was done using JQ which is not
installed on all systems.

Create a Go CLI application which will make a HTTP GET request to Consul
and inspect the results to ensure the status of the specified service
matches the desired state.

Update the Makefile's build section to build the Go application at the
root of the project.

Update the consul-svc-healthy.sh script to use the Go application rather
than curl and JQ.

Signed-off-by: Anthony M. Bonafide <AnthonyMBonafide@gmail.com>